packages/bedrock/pyproject.toml
[build-system]
requires = ["uv_build>=0.9.6,<0.10.0"]
build-backend = "uv_build"

[project]
name = "sb-bedrock"
version = "0.1.0"
description = "Second Brain Bedrock agent runtime and orchestration"
requires-python = ">=3.10"
dependencies = [
    "sb-shared",
    "langchain>=0.0.300",
    "langchain-aws>=0.0.0",
    "bedrock-agentcore",
    "strands-agents==1.24.0",
    "strands-agents-tools==0.2.19",
    "bedrock-agentcore-starter-toolkit",
]

#[tool.uv.sources]
#sb-shared = { workspace }

[tool.uv]
default-groups = ["dev"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "isort>=5.13.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
